de_project/requirements.txt
pandas>=2.0
pyarrow>=14.0
openpyxl>=3.1
azure-storage-blob>=12.19
python-dotenv>=1.0
